+++ lib/components/collapse/collapse.dart
@@ -40,10 +40,10 @@ class BsCollapseDirective {
   /// if `true` the component is shown
   @HostBinding('class.show')
   @HostBinding('attr.aria-expanded')
-  bool expanded = true;
+  bool expanded = false;
 
   @HostBinding('class.collapse')
-  bool collapsed = false;
+  bool collapsed = true;
 
   bool _collapsing = false;
 
@@ -60,6 +60,7 @@ class BsCollapseDirective {
 
   /// sets and fires the collapsed state of the component
   @Input() set bsCollapse(bool value) {
+    print('bsCollapse.value: $value');
     _bsCollapse = value ?? false;
     _bsCollapseChangeController.add(_bsCollapse);
   }
@@ -74,6 +75,10 @@ class BsCollapseDirective {
 
   final _collapsingChangeController = new StreamController<bool>.broadcast();
 
+  Timer showTimer;
+
+  Timer hideTimer;
+
   /// Emits the collapsing state of the component
   @Output() Stream<bool> get collapsingChange =>
       _collapsingChangeController.stream;
@@ -82,9 +87,10 @@ class BsCollapseDirective {
     expanded = false;
     height = _scrollHeight;
     collapsing = true;
+    showTimer?.cancel();
     new Timer(const Duration(milliseconds: 10), () {
       height = '0';
-      new Timer(const Duration(milliseconds: 350), () {
+      hideTimer = new Timer(const Duration(milliseconds: 350), () {
         collapsing = false;
         collapsed = true;
         height = '';
@@ -96,9 +102,10 @@ class BsCollapseDirective {
     collapsed = false;
     height = '0';
     collapsing = true;
-    new Timer(const Duration(milliseconds: 10), () {
+    hideTimer?.cancel();
+    new Future(() {
       height = _scrollHeight;
-      new Timer(const Duration(milliseconds: 350), () {
+      showTimer = new Timer(const Duration(milliseconds: 350), () {
         collapsing = false;
         expanded = true;
         height = '';